+++ BlazorServerKeycloak/Authorization/ClaimType.cs
@@ -1,0 +1,12 @@
+﻿namespace BlazorServerKeycloak.Authorization;
+
+public struct ClaimType
+{
+    public static string AccessToken => "access_token";
+
+    public static string RefreshToken => "refresh_token";
+
+    public static string AccessTokenExpires => "access_token_expires";
+
+    public static string RefreshTokenExpires => "refresh_token_expires";
+}
