--- BlazorServerKeycloak/ClaimType.cs
@@ -1,12 +1,0 @@
-﻿namespace BlazorServerKeycloak;
-
-public struct ClaimType
-{
-    public static string AccessToken => "access_token";
-
-    public static string RefreshToken => "refresh_token";
-
-    public static string AccessTokenExpires => "access_token_expires";
-
-    public static string RefreshTokenExpires => "refresh_token_expires";
-}
