+++ BlazorServerKeycloak/Components/Shared/Login.razor
@@ -1,4 +1,5 @@
 ﻿@using BlazorServerKeycloak
+@using BlazorServerKeycloak.Interfaces
 @using Microsoft.AspNetCore.Components.Authorization
 @using Microsoft.Extensions.DependencyInjection
 @implements IDisposable
