+++ BlazorServerKeycloak/Components/_Imports.razor
@@ -3,4 +3,4 @@
 @using Microsoft.AspNetCore.Components.Forms
 @using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.AspNetCore.Components.Web
-@using BlazorServerKeycloak.Shared
+@using BlazorServerKeycloak.Components.Shared
