--- BlazorServerKeycloak/IUserDisplayGetter.cs
@@ -1,8 +1,0 @@
-﻿using System.Security.Claims;
-
-namespace BlazorServerKeycloak;
-
-public interface IUserDisplayGetter
-{
-    Task<string?> Get(ClaimsPrincipal user);
-}
