+++ BlazorServerKeycloak/Interfaces/IUserDisplayGetter.cs
@@ -1,0 +1,8 @@
+﻿using System.Security.Claims;
+
+namespace BlazorServerKeycloak.Interfaces;
+
+public interface IUserDisplayGetter
+{
+    Task<string?> Get(ClaimsPrincipal user);
+}
